+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Mapbox welcomes participation and contributions from everyone.
 
 * Add a new API to disable custom resizing implementation of the MapView. To disable the custom resizing implementation, set `MapView.resizingAnimation` to `.none`.
 * Add `to-hsla` expression support.
+* Localize geofencing attribution dialog.
 
 ## 11.9.0-beta.1